Properly separate Uri as property and not type
